Change Cat Details to fetch specific breed by id
This was a bit of a change as it impacted how the components received the data in, and had a knock on affect on the tests too. Previously the App was loading all cat breeds and passing as a param into the CatDetails component, but it makes more sense for the page to just get the cat breed required.

Also includes some basic styling of the pages.